+++ components/Input/ModalInputContainer/TagInput.tsx
@@ -1,10 +1,9 @@
 import ContentChip from '@/components/Chip/ContentChip';
 import { StyledLabel } from '@/components/Input/Input.style';
-import { TAG_COLOR } from '@/constants/ModalInput';
+import useTagInput from '@/hooks/useTagInput';
 import { fontStyle } from '@/styles/fontStyle';
 import { COLORS } from '@/styles/palettes';
-
-import { ChangeEvent, KeyboardEvent, MouseEvent, useEffect, useState } from 'react';
+import { MouseEvent } from 'react';
 import styled from 'styled-components';
 
 export interface TagProps {
@@ -25,54 +24,11 @@ export interface Tag {
  * @param label input 라벨 텍스트
  * @param onChange 부모 컴포넌트에서 제어하는 input onChange 함수
  * */
-function TagInput({ label = '태그', onChange, defaultValue: defaultTagStringList }: TagProps) {
-  const [tags, setTags] = useState<Tag[]>([]);
-  const [value, setValue] = useState<string>('');
-
-  const getFormattedTagEl = () => {
-    if (!defaultTagStringList) return;
-    const splitTag = defaultTagStringList.map((val) => val.split('/'));
-    const formattedTag = splitTag.map((val) => ({
-      value: val[0],
-      color: val[1],
-      backgroundColor: val[2],
-    }));
-    return formattedTag;
-  };
-
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setValue(e.target.value);
-  };
-
-  const handleDeleteClick = (selectedValue: string) => {
-    setTags((prev) => prev.filter((prevTags) => prevTags.value !== selectedValue));
-  };
-
-  const handleOnEnterKeyUp = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key !== 'Enter' || value === '') return;
-
-    const ranVal = Math.floor(Math.random() * 4);
-    const TagColors = TAG_COLOR[ranVal];
-    const newTagEl = { value: value, backgroundColor: TagColors[0], color: TagColors[1] };
-
-    setTags((prev) => {
-      if (prev.find((prevTags) => prevTags.value === newTagEl.value)) return prev;
-      return [...prev, newTagEl];
-    });
-
-    setValue('');
-  };
-
-  useEffect(() => {
-    if (defaultTagStringList) {
-      const formatTags = getFormattedTagEl() as Tag[];
-      setTags(formatTags);
-    }
-  }, []);
-
-  useEffect(() => {
-    onChange(label, tags);
-  }, [tags]);
+function TagInput({ label = '태그', onChange, defaultValue }: TagProps) {
+  const { tags, value, handleDeleteClick, handleInputChange, handleOnEnterKeyUp } = useTagInput({
+    defaultValue,
+    onChange,
+  });
 
   return (
     <div>
